Move most asset related code to assets.h
