+++ src/Qubic.vcxproj
@@ -14,6 +14,7 @@
     <ClCompile Include="qubic.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="assets.h" />
     <ClInclude Include="logging.h" />
     <ClInclude Include="network_messages.h" />
     <ClInclude Include="peers.h" />
