+++ src/network_messages.h
@@ -110,6 +110,8 @@ typedef struct
     unsigned char peers[NUMBER_OF_EXCHANGED_PEERS][4];
 } ExchangePublicPeers;
 
+#define END_RESPONSE 35
+
 
 struct TryAgain // Must be returned if _dejavu is not 0, and the incoming packet cannot be processed (usually when incoming packets queue is full)
 {
